Volviendo un componente la vista de edición de atributos
